requirements.txt
accelerate==1.0.1
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
defusedxml==0.7.1
Django==5.1.2
djangorestframework==3.15.2
filelock==3.16.1
fsspec==2024.10.0
greenlet==3.1.1
huggingface-hub==0.26.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
parsel==1.9.1
playwright==1.48.0
Protego==0.3.1
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.0.0
pyOpenSSL==24.2.1
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
safetensors==0.4.5
Scrapy==2.11.2
scrapy-playwright==0.0.41
service-identity==24.1.0
setuptools==75.1.0
sqlparse==0.5.1
sympy==1.13.1
tldextract==5.1.2
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
wheel==0.44.0
zope.interface==7.1.1
